Fixed include bugs and updated build system to include the newest tools.
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@342 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/iWinch/CWinch.cpp
@@ -77,7 +77,7 @@ void CWinch::GetZeroCurrentLevel(void)
 long CWinch::PositionToCounts(double line_pos)
 {
 	if(type == TYPE_LINE) {
-		return (line_pos / line_per_turn) * enc_per_drive_rev;
+		return long((line_pos / line_per_turn) * enc_per_drive_rev);
 	} else if(type == TYPE_SPOOL) {
 		//...
 	}
